# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/rewards:
    get:
      tags:
        - Rewards
      summary: List Rewards
      description: Get a paginated list of Rewards with optional filtering and sorting
      operationId: listRewards
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Rewards
      summary: Create Reward
      description: Create a new Reward
      operationId: createReward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRewardRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reward'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/talent_matrix_entries:
    get:
      tags:
        - TalentMatrixEntries
      summary: List TalentMatrixEntries
      description: Get a paginated list of TalentMatrixEntries with optional filtering and sorting
      operationId: listTalentMatrixEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TalentMatrixEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TalentMatrixEntries
      summary: Create TalentMatrixEntry
      description: Create a new TalentMatrixEntry
      operationId: createTalentMatrixEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTalentMatrixEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TalentMatrixEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Reward:
      type: object
      required:
        - id
        - employee_id
        - reward_type
        - title
        - awarded_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        cycle_id:
          nullable: true
          type: string
          format: uuid
        reward_type:
          type: string
        title:
          type: string
          maxLength: 180
        description:
          nullable: true
          type: string
        amount:
          nullable: true
          type: string
        awarded_by:
          nullable: true
          type: string
          format: uuid
        awarded_at:
          type: string
          format: date
        payroll_component_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateRewardRequest:
      type: object
      required:
        - employee_id
        - reward_type
        - title
        - awarded_at
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        cycle_id:
          nullable: true
          type: string
          format: uuid
        reward_type:
          type: string
        title:
          type: string
          maxLength: 180
        description:
          nullable: true
          type: string
        amount:
          nullable: true
          type: string
        awarded_by:
          nullable: true
          type: string
          format: uuid
        awarded_at:
          type: string
          format: date
        payroll_component_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    RewardList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Reward'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TalentMatrixEntry:
      type: object
      required:
        - id
        - employee_id
        - cycle_id
        - performance_score
        - potential_score
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        cycle_id:
          type: string
          format: uuid
        performance_score:
          type: integer
          format: int32
        potential_score:
          type: integer
          format: int32
        box_label:
          nullable: true
          type: string
          maxLength: 40
        metadata:
          type: object
...
    CreateTalentMatrixEntryRequest:
      type: object
      required:
        - employee_id
        - cycle_id
        - performance_score
        - potential_score
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        cycle_id:
          type: string
          format: uuid
        performance_score:
          type: integer
          format: int32
        potential_score:
          type: integer
          format: int32
        box_label:
          nullable: true
          type: string
          maxLength: 40
        metadata:
          type: object
...
    TalentMatrixEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TalentMatrixEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []